+++ Tower_defence/bullet.h
@@ -12,15 +12,15 @@ public:
     Bullet(Tower* parent, MyPoint target);
     ~Bullet();
     int get_velocity()const{return velocity;}
-    double get_direction_x()const{return direction_x;}
-    double get_direction_y()const{return direction_y;}
+    float get_direction_x()const{return direction_x;}
+    float get_direction_y()const{return direction_y;}
     int getDamage(){return damage;}
 protected:
     Tower* parent;
     int damage;//单颗伤害
     int velocity;//速度
-    double direction_x;
-    double direction_y;//方向
+    float direction_x;
+    float direction_y;//方向
 
 };
