+++ Tower_defence/enemy.h
@@ -10,16 +10,16 @@ public:
     explicit Enemy();
     ~Enemy(){delete icon;}
     int get_velocity()const{return velocity;}
-    double get_direction_x()const{return direction_x;}
-    double get_direction_y()const{return direction_y;}
+    float get_direction_x()const{return direction_x;}
+    float get_direction_y()const{return direction_y;}
     int getLife(){return life;}
     void setLife(int l){life = l;}
 
 protected:
     int life;
     int velocity;
-    double direction_x;
-    double direction_y;
+    float direction_x;
+    float direction_y;
 
 };
