+++ Tower_defence/tool.cpp
@@ -7,15 +7,15 @@ Tool::Tool()
 {
 
 }
-double Tool::vector_to_angle(MyPoint p){
-    double x = p.x();
-    double y = p.y();
-    double bevel = pow(pow(x,2)+pow(y,2), 0.5);//斜边
-    double _sin = y/bevel;
-    double _cos = x/bevel;
-    double ans1 = asin(_sin);
-    double ans2 = acos(_cos);
-    double ans;
+float Tool::vector_to_angle(MyPoint p){
+    float x = p.x();
+    float y = p.y();
+    float bevel = pow(pow(x,2)+pow(y,2), 0.5);//斜边
+    float _sin = y/bevel;
+    float _cos = x/bevel;
+    float ans1 = asin(_sin);
+    float ans2 = acos(_cos);
+    float ans;
     if(_sin >= 0 && _cos >= 0){
         ans = ans1;
     }else if(_sin >= 0 && _cos <= 0){
@@ -27,14 +27,14 @@ double Tool::vector_to_angle(MyPoint p){
     }
     return ans*180/M_PI;
 }
-MyPoint Tool::angel_to_vector(double angle){
-    double ang = angle/180*M_PI;
-    double _sin = sin(ang);
-    double _cos = cos(ang);
+MyPoint Tool::angel_to_vector(float angle){
+    float ang = angle/180*M_PI;
+    float _sin = sin(ang);
+    float _cos = cos(ang);
     MyPoint p;
     p.setX(_cos);
     p.setY(_sin);
     return p;
 }
-int Tool::sign(double a){return a>0?1:-1;}
-double Tool::abs(double a){return a>0?a:-a;}
+int Tool::sign(float a){return a>0?1:-1;}
+float Tool::abs(float a){return a>0?a:-a;}
